refactor: move index.js and standalone.js to src
